Move check_declination
