+++ dolfyn/rotate/base.py
@@ -1,5 +1,6 @@
 import numpy as np
 from numpy.linalg import det
+from .base_nortek import orient2euler, euler2orient
 
 
 class BadDeterminantWarning(UserWarning):
@@ -54,3 +55,54 @@ def _check_rotmat_det(rotmat, thresh=1e-3):
     return np.abs(det(rotmat) - 1) < thresh
 
 
+def _check_declination(advo):
+    odata = advo['orient']
+    cs = advo.props['coord_sys']
+    if 'declination' not in advo.props:
+        if 'orientmat' in odata and \
+           advo._make_model.startswith('nortek vector'):
+            # Vector's don't have p,r,h when they have an orientmat.
+            p, r, h = orient2euler(odata['orientmat'])
+            odata['pitch'] = p
+            odata['roll'] = r
+            odata['heading'] = h
+        # warnings.warn(
+        #     'No declination in adv object.  Assuming a declination of 0.')
+        return
+
+    if 'orientmat' in odata and \
+       not advo.props.get('declination_in_orientmat', False):
+        # Declination is defined as positive if MagN is east of
+        # TrueN. Therefore we must rotate about the z-axis by minus
+        # the declination angle to get from Mag to True.
+        cd = np.cos(-np.deg2rad(advo.props['declination']))
+        sd = np.sin(-np.deg2rad(advo.props['declination']))
+        # The ordering is funny here because orientmat is the
+        # transpose of the inst->earth rotation matrix:
+        Rdec = np.array([[cd, -sd, 0],
+                         [sd, cd, 0],
+                         [0, 0, 1]])
+        odata['orientmat'] = np.einsum('ij,kjl->kil',
+                                       Rdec,
+                                       odata['orientmat'])
+        # NOTE: for a moment I thought I needed to do a tensor
+        # rotation on orientmat, but that's not the case. It seems
+        # like the rotation matrix isn't actually a "tensor".
+        # I checked this by showing that the above actually gives the
+        # desired result of rotating vectors by the declination.
+
+        advo.props['declination_in_orientmat'] = True
+        if advo._make_model.startswith('nortek vector'):
+            p, r, h = orient2euler(odata['orientmat'])
+            odata['pitch'] = p
+            odata['roll'] = r
+            odata['heading'] = h
+            advo.props['declination_in_heading'] = True
+
+    if 'heading' in odata and \
+       not advo.props.get('declination_in_heading', False):
+
+        odata['heading'] += advo.props['declination']
+        odata['heading'][odata['heading'] < 0] += 360
+        odata['heading'][odata['heading'] > 360] -= 360
+        advo.props['declination_in_heading'] = True
