+++ dolfyn/rotate/signature.py
@@ -1,7 +1,7 @@
 from . import base as rotb
 from .vector import earth2principal
 from .rdi import beam2inst
-from .base_nortek import _check_declination, euler2orient
+from .base import _check_declination, euler2orient
 import numpy as np
 import warnings
 from numpy.linalg import inv
