+++ dolfyn/rotate/vector.py
@@ -2,7 +2,7 @@ from __future__ import division
 import numpy as np
 import warnings
 from numpy.linalg import inv
-from .base_nortek import _check_declination, euler2orient, orient2euler
+from .base import _check_declination, euler2orient, orient2euler
 from . import base as rotb
